- Refactorisation d'une partie du code pour plus de simplicité.
- Correction d'un bug sur le nombre de succès auto de l'immobilisation du mode morph.
- Correction d'un bug empêchant d'ajouter des effets sur une capacité de PNJ / Créature si elle est sur un token.
- Correction d'un bug d'affichage lorsque l'ont permet de modifier le nombre de dés des nods, sur les PNJ.

+++ module/utils/socketHandler.mjs
@@ -26,6 +26,8 @@ async function giveItmToActor(payload={}) {
     const actor = await fromUuid(payload.actor);
     const items = payload.items;
 
+    console.error(items);
+
     actor.createEmbeddedDocuments("Item", items);
 }
 
@@ -41,12 +43,14 @@ async function createToken(payload={}) {
     const data = payload.data;
 
     const created = await scene.createEmbeddedDocuments("Token", data);
+
     return created?.[0];
 }
 
 async function deleteActorOrToken(payload={}) {
     for(let a of payload.actors) {
         const actor = await fromUuid(a);
+        if(!actor) continue;
 
         actor.delete();
     }
